add auth middleware, userschema, role support

+++ routes/products.js
@@ -2,8 +2,9 @@ var express = require('express')
 var prodroute = express.Router()
 let product = require('../models/product');
 let category = require('../models/category');
+const auth = require("../middleware/auth");
 
-prodroute.get('/:category_id', async function (req, res) {
+prodroute.get('/:category_id', auth, async function (req, res) {
     try {
         if (req.params.category_id == "all") {
             products = await product.find({})
@@ -19,35 +20,47 @@ prodroute.get('/:category_id', async function (req, res) {
     }
 })
 
-prodroute.post('/add', async function (req, res) {
+prodroute.post('/add', auth, async function (req, res) {
     try {
         // only by Supervisor
-        newprod = new product(req.body)
-        result = await newprod.save()
-        console.log(result)
-        res.status(200).send('Add a product to category')
+        if (req.user.role == "Supervisor") {
+            newprod = new product(req.body)
+            result = await newprod.save()
+            console.log(result)
+            res.status(200).send('Added product successfully')
+        } else {
+            res.status(400).send(`You cant add product, role: ${req.user.role}`)
+        }
     } catch (error) {
         res.status(401).send(`Error while inserting product: ${error}`)
     }
 })
 
-prodroute.post('/update', async function (req, res) {
+prodroute.post('/update', auth, async function (req, res) {
     // only by Supervisor
     try {
-        newprod = new product(req.body.data);
-        await product.updateOne({ _id: req.body._id }, req.body.data, { runValidators: true })
-        res.status(200).send('update prod details')
+        if (req.user.role == "Supervisor") {
+            newprod = new product(req.body.data);
+            await product.updateOne({ _id: req.body._id }, req.body.data, { runValidators: true })
+            res.status(200).send('update prod details')
+        } else {
+            res.status(400).send(`You cant update product, role: ${req.user.role}`)
+        }
     } catch (error) {
         res.status(400).send(`Error while updating prod: ${error}`)
     }
 })
 
-prodroute.delete('/:id', async function (req, res) {
+prodroute.delete('/:id', auth, async function (req, res) {
     // only by Supervisor
     try {
-        result = await product.deleteOne({ _id: req.params.id })
-        console.log(result)
-        res.send(`delete prod details: ${JSON.stringify(result)}`)
+        if (req.user.role == "Supervisor") {
+            result = await product.deleteOne({ _id: req.params.id })
+            console.log(result)
+            res.send(`delete prod details: ${JSON.stringify(result)}`)
+        } else {
+            res.status(400).send(`You cant delete product, role: ${req.user.role}`)
+        }
     } catch (error) {
         res.status(400).send(`Error while deleting: ${error}`)
     }
